+++ src/org/xjge/core/TCListSpeakers.java
@@ -23,15 +23,15 @@ final class TCListSpeakers extends TerminalCommand {
     public TerminalOutput execute(List<String> args) {
         var speakerList = new StringBuilder();
         
-        for(int i = 0; i < Audio.getNumSpeakers(); i++) {
-            var speaker = Audio.getSpeaker(i);
+        for(int i = 0; i < Audio2.getNumSpeakers(); i++) {
+            var speaker = Audio2.getSpeaker(i);
             
             speakerList.append("\"").append(speaker.name).append("\" at index ").append(i);
             
-            if(i < Audio.getNumSpeakers() - 1) speakerList.append(" \n");
+            if(i < Audio2.getNumSpeakers() - 1) speakerList.append(" \n");
         }
         
-        return new TerminalOutput(speakerList.toString(), Color.WHITE);
+        return new TerminalOutput(speakerList.toString(), Color.YELLOW);
     }
 
 }